+++ src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
+import { ApplicationModule } from '@/application/application.module';
 import { envSchema } from '@/config/env';
 import { PrismaModule } from '@/database/prisma/prisma.module';
-import { ApplicationModule } from './application/application.module';
 
 @Module({
 	imports: [
