+++ src/infra/http/controllers/delete-user.controller.ts
@@ -1,3 +1,4 @@
+import type { AuthenticatedUserRequest } from '@dtos/get-user.dto';
 import {
 	Controller,
 	Delete,
@@ -12,11 +13,10 @@ import {
 	ApiOperation,
 	ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
+import { UserService } from '@services/user.service';
 import { ROLE } from '@/_types';
-import { UserService } from '@/application/services/user.service';
-import { Roles } from '@/modules/auth/decorators/roles.decorator';
-import { JwtAuthGuard } from '@/modules/auth/guards/jwt-auth.guard';
-import type { AuthenticatedUserRequest } from '../dto/get-user.dto';
+import { Roles } from '@/shared/decorators/roles.decorator';
+import { JwtAuthGuard } from '../../../shared/guards/jwt-auth.guard';
 
 @Controller('/accounts')
 export class DeleteUserController {
